[integration_test] split the vllm e2e harness and adapt it to the core split
e2e_lib.py (839 lines) carried everything; it is now three modules with
one concern each, behind a re-exporting facade so scenario files keep
their imports:

* lib_utils.py -- pure helpers (model detection, paths, tokenization,
  manager queries, traffic, capture comparison); no process state;
* servers.py -- process lifecycle: the real manager binary, the real
  vLLM OpenAI server and ScenarioEnv;
* e2e_lib.py -- compatibility facade only.

Review follow-ups baked in:

* the hybrid model detection prefers declared architectures
  (Qwen3Next/Zamba/FalconH1/Samba/Jamba markers) and falls back to the
  layer-knob heuristics only when the config declares none;
* the vLLM serving knobs (max-model-len, gpu-memory-utilization, ...)
  move into an override-able vllm_args dict instead of hardcoded CLI
  fragments -- a scenario can widen them without editing the harness;
* the VLLM_KV_CACHE_LAYOUT=NHD comment now answers the question asked
  in review: it is required, not cosmetic -- the connector asserts
  token-major pages at register_kv_caches and refuses to start on an
  HND-default vLLM without it.

test_connector.py adapts to the connector's new shape: the worker keeps
no request mirror any more, so the VerifyingConnector's scheduler side
snapshots the ledger's live token streams onto the envelope
(token_snapshots) for the worker-side captures; the mutated meta-test
swaps connector_worker instead of _core; captures dispatch on
AttentionGroupMeta and the (views, layout) attn_kv_views signature.
